requirements.txt
pyyaml
discord.py
python-dotenv
